+++ lib/src/matrix/matrix_cache_keys.dart
@@ -4,6 +4,7 @@ const matrixDeviationByRowsKey = 'deviation_by_rows';
 const matrixDeviationByColumnsKey = 'deviation_by_columns';
 const matrixPowKey = 'pow';
 const matrixExpKey = 'exp';
+const matrixLogKey = 'log';
 const matrixSumKey = 'sum';
 const matrixProdKey = 'prod';
 
@@ -14,6 +15,7 @@ final matrixCacheKeys = Set<String>.from(<String>[
   matrixDeviationByColumnsKey,
   matrixPowKey,
   matrixExpKey,
+  matrixLogKey,
   matrixSumKey,
   matrixProdKey,
 ]);