+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 12.17.0
+- `Vector`:
+    - `Vector.log` method added
+- `Matrix`:
+    - `Matrix.log` method added
+
 ## 12.16.0
 - `Matrix`:
     - `Matrix.mapElements` method added